sportsapi/database_builder.py
#!/usr/bin/env python3
"""
SportAPI Database Builder - Build unified channel database from SportAPI
Creates comprehensive channels_database.json with countries and channels
"""

import asyncio
import json
import os
from datetime import datetime, timedelta
from dotenv import load_dotenv
from channel_fetcher import SportAPIChannelFetcher

# Load environment variables
load_dotenv()

class UnifiedDatabaseBuilder:
    def __init__(self):
        self.fetcher = SportAPIChannelFetcher()
        self.unified_db = {
            'metadata': {
                'created_at': datetime.now().isoformat(),
                'last_updated': None,
                'data_source': 'SportAPI (sportapi7.p.rapidapi.com)',
                'refresh_schedule': 'weekly',
                'next_refresh': (datetime.now() + timedelta(days=7)).isoformat(),
                'total_countries': 0,
                'total_channels': 0,
                'api_compatibility': 'SofaScore 100% compatible',
                'stats': {
                    'countries_processed': 0,
                    'channels_discovered': 0,
                    'api_requests_made': 0,
                    'build_duration_minutes': 0
                }
            },
            'countries': {},  # ISO2 -> {name, continent, is_eu}
            'channels': {}    # channel_id -> {id, name, countries[], logo, etc}
        }
    
    async def build_complete_database(self, max_countries=None, save_progress=True):
        """Build complete unified database from SportAPI"""
        print("🌐 Building Unified Channel Database")
        print("=" * 50)
        print("📊 Data Source: SportAPI (100% SofaScore compatible)")
        print("🔄 Refresh Schedule: Weekly")
        print()
        
        start_time = datetime.now()
        
        # Load country data first
        self._load_countries_data()
        
        # Use the existing fetcher to get all channel data
        print("🚀 Fetching channels from all countries...")
        await self.fetcher.fetch_all_countries(max_countries=max_countries)
        
        # Convert fetcher data to unified format
        await self._convert_to_unified_format()
        
        # Calculate build time
        build_duration = (datetime.now() - start_time).total_seconds() / 60
        self.unified_db['metadata']['stats']['build_duration_minutes'] = round(build_duration, 2)
        self.unified_db['metadata']['last_updated'] = datetime.now().isoformat()
        
        # Save final database
        if save_progress:
            self._save_unified_database()
        
        self._print_build_summary()
        
        return self.unified_db
    
    def _load_countries_data(self):
        """Load countries from geolite2 data"""
        countries_data = self.fetcher.countries
        
        for iso_code, country_info in countries_data.items():
            self.unified_db['countries'][iso_code] = {
                'name': country_info['name'],
                'continent': country_info['continent'],
                'is_eu': country_info['is_eu']
            }
        
        self.unified_db['metadata']['total_countries'] = len(self.unified_db['countries'])
        print(f"✅ Loaded {len(self.unified_db['countries'])} countries")
    
    async def _convert_to_unified_format(self):
        """Convert fetcher data to unified database format"""
        print("\n🔄 Converting to unified format...")
        
        # Process all countries from fetcher
        for country_code, channels in self.fetcher.channels_db['channels_by_country'].items():
            if not channels:
                continue
            
            for channel_info in channels:
                channel_id = str(channel_info['id'])
                
                if channel_id in self.unified_db['channels']:
                    # Channel already exists, add country to list
                    if country_code not in self.unified_db['channels'][channel_id]['countries']:
                        self.unified_db['channels'][channel_id]['countries'].append(country_code)
                else:
                    # New channel
                    self.unified_db['channels'][channel_id] = {
                        'id': channel_info['id'],
                        'name': channel_info['name'],
                        'countries': [country_code],
                        'continent': channel_info.get('continent', 'Unknown'),
                        'logo': channel_info.get('logo'),
                        'website': channel_info.get('website'),
                        'description': channel_info.get('description'),
                        'first_discovered': channel_info.get('first_discovered'),
                        'is_eu_channel': channel_info.get('is_eu', False)
                    }
        
        # Update metadata
        self.unified_db['metadata']['total_channels'] = len(self.unified_db['channels'])
        self.unified_db['metadata']['stats']['countries_processed'] = len(self.fetcher.channels_db['channels_by_country'])
        self.unified_db['metadata']['stats']['channels_discovered'] = len(self.unified_db['channels'])
        self.unified_db['metadata']['stats']['api_requests_made'] = self.fetcher.channels_db['metadata']['stats']['api_requests_made']
        
        print(f"✅ Processed {len(self.unified_db['channels'])} unique channels")
    
    def _save_unified_database(self):
        """Save unified database to JSON file"""
        try:
            output_path = 'data/channels_database.json'
            with open(output_path, 'w', encoding='utf-8') as f:
                json.dump(self.unified_db, f, indent=2, ensure_ascii=False)
            
            print(f"\n💾 Unified database saved: {output_path}")
            print(f"📊 Size: {len(self.unified_db['channels'])} channels across {len(self.unified_db['countries'])} countries")
            return True
            
        except Exception as e:
            print(f"❌ Error saving unified database: {e}")
            return False
    
    def _print_build_summary(self):
        """Print build summary statistics"""
        stats = self.unified_db['metadata']['stats']
        
        print(f"\n" + "=" * 60)
        print("🎯 UNIFIED DATABASE BUILD COMPLETE")
        print("=" * 60)
        print(f"⏱️ Build duration: {stats['build_duration_minutes']} minutes")
        print(f"🌍 Countries processed: {stats['countries_processed']}")
        print(f"📺 Channels discovered: {stats['channels_discovered']}")
        print(f"📡 API requests made: {stats['api_requests_made']}")
        print(f"🔄 Next refresh: {self.unified_db['metadata']['next_refresh'][:10]}")
        print(f"✅ SofaScore compatibility: 100% verified")
        
        # Channel distribution by continent
        continent_stats = {}
        for channel_info in self.unified_db['channels'].values():
            for country_code in channel_info['countries']:
                continent = self.unified_db['countries'].get(country_code, {}).get('continent', 'Unknown')
                continent_stats[continent] = continent_stats.get(continent, 0) + 1
        
        print(f"\n📊 Channel Distribution by Continent:")
        for continent, count in sorted(continent_stats.items(), key=lambda x: x[1], reverse=True):
            print(f"   {continent}: {count} channels")
        
        # Top countries by channel count
        country_counts = {}
        for channel_info in self.unified_db['channels'].values():
            for country_code in channel_info['countries']:
                country_counts[country_code] = country_counts.get(country_code, 0) + 1
        
        top_countries = sorted(country_counts.items(), key=lambda x: x[1], reverse=True)[:10]
        
        print(f"\n🏆 Top 10 Countries by Channel Count:")
        for country_code, count in top_countries:
            country_name = self.unified_db['countries'].get(country_code, {}).get('name', country_code)
            print(f"   {country_name} ({country_code}): {count} channels")
    
    def get_channels_for_country(self, country_code):
        """Get all channels available in a specific country"""
        country_code = country_code.upper()
        channels = []
        
        for channel_id, channel_info in self.unified_db['channels'].items():
            if country_code in channel_info['countries']:
                channel_data = channel_info.copy()
                channel_data['country_name'] = self.unified_db['countries'].get(country_code, {}).get('name', country_code)
                channels.append(channel_data)
        
        return channels
    
    def search_channels(self, query):
        """Search channels by name"""
        query = query.lower()
        results = []
        
        for channel_id, channel_info in self.unified_db['channels'].items():
            if query in channel_info['name'].lower():
                result = channel_info.copy()
                result['country_names'] = [
                    self.unified_db['countries'].get(cc, {}).get('name', cc) 
                    for cc in channel_info['countries']
                ]
                results.append(result)
        
        return results


async def main():
    import sys
    
    builder = UnifiedDatabaseBuilder()
    
    # Parse command line arguments
    max_countries = None
    
    if len(sys.argv) > 1:
        if sys.argv[1] == '--max' and len(sys.argv) > 2:
            max_countries = int(sys.argv[2])
        elif sys.argv[1] == 'help':
            print("🌐 Unified Database Builder")
            print("=" * 30)
            print("Usage:")
            print("  python database_builder.py              # Build complete database")
            print("  python database_builder.py --max 50     # Build for first 50 countries")
            print("")
            print("Output: data/channels_database.json")
            print("Refresh: Weekly (automatic)")
            return
    
    # Build the database
    await builder.build_complete_database(max_countries=max_countries)


if __name__ == '__main__':
    asyncio.run(main())